requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.6
    # via fastmcp
beartype==0.22.9
    # via
    #   py-key-value-aio
    #   py-key-value-shared
boto3==1.42.17
    # via -r requirements.in
botocore==1.42.17
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.4
    # via py-key-value-aio
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
cloudpickle==3.1.2
    # via pydocket
colorama==0.4.6
    # via click
cryptography==46.0.3
    # via
    #   authlib
    #   pyjwt
cyclopts==4.4.1
    # via fastmcp
dataclasses==0.6
    # via -r requirements.in
diskcache==5.6.3
    # via py-key-value-aio
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.4
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.1
    # via fastmcp
fakeredis[lua]==2.33.0
    # via pydocket
fastmcp==2.14.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.2
    # via keyring
jaraco-functools==4.4.0
    # via keyring
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.25.1
    # via mcp
jsonschema-path==0.3.4
    # via fastmcp
jsonschema-specifications==2025.9.1
    # via jsonschema
keyring==25.7.0
    # via py-key-value-aio
lupa==2.6
    # via fakeredis
markdown-it-py==4.0.0
    # via rich
mcp==1.25.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via
    #   jaraco-classes
    #   jaraco-functools
openapi-pydantic==0.5.1
    # via fastmcp
opentelemetry-api==1.39.1
    # via
    #   opentelemetry-exporter-prometheus
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydocket
opentelemetry-exporter-prometheus==0.60b1
    # via pydocket
opentelemetry-instrumentation==0.60b1
    # via pydocket
opentelemetry-sdk==1.39.1
    # via opentelemetry-exporter-prometheus
opentelemetry-semantic-conventions==0.60b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
pathable==0.4.4
    # via jsonschema-path
pathvalidate==3.3.1
    # via py-key-value-aio
pillow==12.0.0
    # via -r requirements.in
platformdirs==4.5.1
    # via fastmcp
prometheus-client==0.23.1
    # via
    #   opentelemetry-exporter-prometheus
    #   pydocket
py-key-value-aio[disk,keyring,memory,redis]==0.3.0
    # via
    #   fastmcp
    #   pydocket
py-key-value-shared==0.3.0
    # via py-key-value-aio
pycparser==2.23
    # via cffi
pydantic[email]==2.12.5
    # via
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pydocket==0.16.3
    # via fastmcp
pygments==2.19.2
    # via rich
pyjwt[crypto]==2.10.1
    # via mcp
pyperclip==1.11.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   fastmcp
    #   pydantic-settings
python-json-logger==4.0.0
    # via pydocket
python-multipart==0.0.21
    # via mcp
pywin32==311
    # via mcp
pywin32-ctypes==0.2.3
    # via keyring
pyyaml==6.0.3
    # via jsonschema-path
redis==7.1.0
    # via
    #   fakeredis
    #   py-key-value-aio
    #   pydocket
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   jsonschema-path
rich==14.2.0
    # via
    #   cyclopts
    #   fastmcp
    #   pydocket
    #   rich-rst
    #   typer
rich-rst==1.3.2
    # via cyclopts
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via fakeredis
sse-starlette==3.1.1
    # via mcp
starlette==0.50.0
    # via
    #   mcp
    #   sse-starlette
typer==0.21.0
    # via pydocket
typing-extensions==4.15.0
    # via
    #   mcp
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   py-key-value-shared
    #   pydantic
    #   pydantic-core
    #   pydocket
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.2
    # via
    #   botocore
    #   requests
uvicorn==0.40.0
    # via
    #   fastmcp
    #   mcp
websockets==15.0.1
    # via fastmcp
wrapt==1.17.3
    # via opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
